requirements.txt
Pillow
PyPDF2
fontTools